Test if $facet is set before deleting
Add testing to make sure we only attempt to delete a facet which is actually set.

+++ SolrSearchPlugin.php
@@ -265,7 +265,10 @@ SQL
     {
         $table = $this->_db->getTable('SolrSearchField');
         $facet = $table->findByElement($args['record']);
-        $facet->delete();
+        
+        if(!empty($facet)) {
+            $facet->delete();
+        }
     }
